requirements.txt
logging == 0.4.9.6
python-telegram-bot == 13.7
beautifulsoup4 == 4.9.3
requests == 2.26.0